% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickns-package.R
\docType{package}
\name{quickns-package}
\alias{quickns}
\alias{quickns-package}
\title{quickns: A package for faster natural spline}
\description{
This package provides a function \code{quick_ns} that can generate natural cubic spline basis faster.
}
\details{
This package provides a faster natural cubic spline basis function, \code{\link{quick_ns}}.
}
\references{
R Core Team (2020). R: A language and environment for statistical computing. R Foundation for Statistical Computing, Vienna, Austria. URL https://www.R-project.org/.
}
\seealso{
\code{\link[splines]{ns}}
}
\author{
Weidong Wang, \email{wang@umass.edu}
}
